refactor(micronauts): bump up micronauts version
