+++ gradle.properties
@@ -2,4 +2,4 @@
 version=3.0.0
 versioneye.projectkey=maven2_master_server_1
 versioneye.projectid=55352d6edc39815abf0007ad
-micronautVersion=2.0.3
+micronautVersion=3.7.2
