+++ src/main/java/org/terasology/web/services/impl/geo/dbip/DbIpGeoLocationService.java
@@ -17,10 +17,10 @@
 package org.terasology.web.services.impl.geo.dbip;
 
 import io.micronaut.context.annotation.Value;
+import jakarta.inject.Singleton;
 import org.terasology.web.services.api.GeoLocationService;
 import org.terasology.web.services.impl.geo.GeoLocation;
 
-import javax.inject.Singleton;
 import java.io.IOException;
 import java.net.InetAddress;
