+++ src/test/java/org/terasology/master/JooqDbTest.java
@@ -16,12 +16,12 @@
 
 package org.terasology.master;
 
-import io.micronaut.test.annotation.MicronautTest;
+import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
+import jakarta.inject.Inject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.terasology.web.services.api.DatabaseService;
 
-import javax.inject.Inject;
 import java.util.Map;
 
 @MicronautTest
