+++ src/test/java/org/terasology/master/ModuleJsonListTest.java
@@ -22,7 +22,8 @@ import io.micronaut.http.HttpStatus;
 import io.micronaut.http.client.HttpClient;
 import io.micronaut.http.client.annotation.Client;
 import io.micronaut.http.client.exceptions.HttpClientResponseException;
-import io.micronaut.test.annotation.MicronautTest;
+import io.micronaut.test.extensions.junit5.annotation.MicronautTest;
+import jakarta.inject.Inject;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.terasology.module.ModuleMetadata;
@@ -30,7 +31,6 @@ import org.terasology.module.ModuleMetadataJsonAdapter;
 import org.terasology.naming.Name;
 import org.terasology.naming.Version;
 
-import javax.inject.Inject;
 import java.io.IOException;
 import java.io.Reader;
 import java.io.StringReader;
